+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CqlExecutorImpl.java
@@ -18,18 +18,26 @@
 package com.palantir.atlasdb.keyvalue.cassandra;
 
 import java.net.InetSocketAddress;
+import java.nio.ByteBuffer;
+import java.nio.charset.StandardCharsets;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import org.apache.cassandra.thrift.Compression;
 import org.apache.cassandra.thrift.ConsistencyLevel;
+import org.apache.cassandra.thrift.CqlPreparedResult;
 import org.apache.cassandra.thrift.CqlResult;
 import org.apache.cassandra.thrift.CqlRow;
 import org.apache.cassandra.thrift.UnavailableException;
 import org.apache.thrift.TException;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.InsufficientConsistencyException;
@@ -47,6 +55,8 @@ public class CqlExecutorImpl implements CqlExecutor {
 
     public interface QueryExecutor {
         CqlResult execute(CqlQuery cqlQuery, byte[] rowHintForHostSelection);
+        CqlPreparedResult prepare(ByteBuffer query, byte[] rowHintForHostSelection, Compression compression);
+        CqlResult executePrepared(int queryId, List<ByteBuffer> values);
     }
 
     CqlExecutorImpl(CassandraClientPool clientPool, ConsistencyLevel consistency) {
@@ -63,15 +73,29 @@ public class CqlExecutorImpl implements CqlExecutor {
             TableReference tableRef,
             List<byte[]> rowsAscending,
             int limit) {
-        String selQuery = "SELECT key, column1, column2 FROM %s WHERE key IN (%s) LIMIT %s;";
-        CqlQuery query = new CqlQuery(selQuery, quotedTableName(tableRef), keys(rowsAscending), limit(limit));
-        return executeAndGetCells(query, rowsAscending.get(0), CqlExecutorImpl::getCellFromRow);
-    }
+        String preparedSelQuery = String.format("SELECT key, column1, column2 FROM %s WHERE key = ? LIMIT %d;",
+                quotedTableName(tableRef).getValue(),
+                limit);
+        ByteBuffer queryBytes = ByteBuffer.wrap(preparedSelQuery.getBytes(StandardCharsets.UTF_8));
 
-    private Arg<String> keys(List<byte[]> rowsAscending) {
-        return UnsafeArg.of("keys",
-                rowsAscending.stream().map(CqlExecutorImpl::getKey)
-                        .collect(Collectors.joining(",")));
+        CqlPreparedResult preparedResult = queryExecutor.prepare(queryBytes, rowsAscending.get(0), Compression.NONE);
+        int queryId = preparedResult.getItemId();
+
+        List<CellWithTimestamp> result = Lists.newArrayList();
+        Iterator<byte[]> rows = rowsAscending.iterator();
+        while (result.size() < limit && rows.hasNext()) {
+            // We can parallelise this, but would need to collect it carefully to make sure we satisfy the guarantees:
+            // 1. Results are ordered
+            // 2. There are at most <limit> results
+            // 3. We return the first <limit> results
+            // We can ensure this by not applying the limit until after collection, and grouping by row
+            // I think this is ensured later on in the code path?
+            CqlResult cqlResult = queryExecutor.executePrepared(queryId,
+                    ImmutableList.of(ByteBuffer.wrap(rows.next())));
+            result.addAll(CqlExecutorImpl.getCells(CqlExecutorImpl::getCellFromRow, cqlResult));
+        }
+
+        return result;
     }
 
     /**
@@ -166,9 +190,12 @@ public class CqlExecutorImpl implements CqlExecutor {
         private final CassandraClientPool clientPool;
         private final ConsistencyLevel consistency;
 
+        private Map<Integer, InetSocketAddress> hostsPerPreparedQuery;
+
         QueryExecutorImpl(CassandraClientPool clientPool, ConsistencyLevel consistency) {
             this.clientPool = clientPool;
             this.consistency = consistency;
+            this.hostsPerPreparedQuery = Maps.newHashMap();
         }
 
         @Override
@@ -176,13 +203,44 @@ public class CqlExecutorImpl implements CqlExecutor {
             return executeQueryOnHost(cqlQuery, getHostForRow(rowHintForHostSelection));
         }
 
+        @Override
+        public CqlPreparedResult prepare(ByteBuffer query, byte[] rowHintForHostSelection, Compression compression) {
+            FunctionCheckedException<CassandraClient, CqlPreparedResult, TException> prepareFunction = client ->
+                    client.rawClient().prepare_cql3_query(query, compression);
+
+            try {
+                InetSocketAddress hostForRow = getHostForRow(rowHintForHostSelection);
+                CqlPreparedResult preparedResult = clientPool.runWithRetryOnHost(hostForRow,
+                        prepareFunction);
+                hostsPerPreparedQuery.put(preparedResult.getItemId(), hostForRow);
+                return preparedResult;
+            } catch (TException e) {
+                throw Throwables.throwUncheckedException(e);
+            }
+        }
+
+        @Override
+        public CqlResult executePrepared(int queryId, List<ByteBuffer> values) {
+            FunctionCheckedException<CassandraClient, CqlResult, TException> cqlFunction = client ->
+                    client.rawClient().execute_prepared_cql3_query(queryId, values, consistency);
+
+            InetSocketAddress host = hostsPerPreparedQuery.getOrDefault(queryId, getHostForRow(values.get(0).array()));
+
+            return executeFunctionOnHost(cqlFunction, host);
+        }
+
         private InetSocketAddress getHostForRow(byte[] row) {
             return clientPool.getRandomHostForKey(row);
         }
 
         private CqlResult executeQueryOnHost(CqlQuery cqlQuery, InetSocketAddress host) {
+            return executeFunctionOnHost(createCqlFunction(cqlQuery), host);
+        }
+
+        private CqlResult executeFunctionOnHost(
+                FunctionCheckedException<CassandraClient, CqlResult, TException> cqlFunction, InetSocketAddress host) {
             try {
-                return clientPool.runWithRetryOnHost(host, createCqlFunction(cqlQuery));
+                return clientPool.runWithRetryOnHost(host, cqlFunction);
             } catch (UnavailableException e) {
                 throw wrapIfConsistencyAll(e);
             } catch (TException e) {